Fix issue with spinner never stopping (#26)
I was experiencing [the same issue as reported here](https://github.com/quasarframework/quasar/issues/888), and learned that the slot="message" attribute needs to be moved upwards to the highest div parent of the spinner, for it to work as expected. The slot="message" attribute should only be part of the q-spinner tag if it is not enclosed in a parent div.

Would be nice to have this also reflected here in the play demo/documentation.